refactor: mvc 기반 패키지 분류
