+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import java.net.URI;
 import java.util.List;
@@ -9,6 +9,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.dto.ReservationCreateDto;
+import roomescape.model.Reservation;
+import roomescape.model.Reservations;
 
 @Controller
 public class RoomescapeController {
