+++ src/main/java/roomescape/model/Reservations.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.model;
 
 import java.util.List;
 import java.util.Map;
